Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.32.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
